splash,cmd: update to handle units consistently
This commit fixes a unit mixup that caused confusion. Previously the main program
was responsible for providing a rectangle, and this rectangle was used as the bounding
box of the progress bar. However, this rectangle was in screen pixels, and thus wasn't
being scaled using gtx.Dp. It would be the same number of physical pixels no matter the
system scaling factor.

This commit restructures the splash widget to accept the height and inset of the progress
bar in Dp, and the splash widget itself handles converting these to screen pixels during
layout. Delaying this unit conversion until the point of layout ensures that even if the
window changes scaling factors, the UI will respond appropriately on the next frame.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>